Rename and refactor poker module files
Renamed poker.py to poker_game.py and split classes into separate files for better organization. Introduced new files for PokerSettings, PokerHandPot, and PokerAction and revised imports and definitions accordingly.
