+++ core/poker_action.py
@@ -1,0 +1,52 @@
+from enum import Enum
+from typing import Optional
+
+
+class PlayerAction(Enum):
+    FOLD = "fold"
+    CHECK = "check"
+    CALL = "call"
+    BET = "bet"
+    RAISE = "raise"
+    ALL_IN = "all-in"
+    NONE = None  # TODO: remove None as an option
+
+
+# TODO: write unit tests for PokerAction
+class PokerAction:
+    player: str
+    player_action: PlayerAction
+    amount: Optional[int]
+    hand_state: Optional[dict]
+    action_detail: Optional[str]
+
+    def __init__(self,
+                 player: str,
+                 action: str,
+                 amount: int or None = None,
+                 hand_state: dict or None = None,
+                 action_detail: str or None = None):
+        self.player = player
+        self.player_action = PlayerAction(action)
+        self.amount = amount
+        self.hand_state = hand_state.copy()
+        self.action_detail = action_detail
+
+    def to_dict(self):
+        return {
+            'player': self.player,
+            'player_action': self.player_action,
+            'amount': self.amount,
+            'hand_state': self.hand_state,
+            'action_detail': self.action_detail
+        }
+
+    @classmethod
+    def from_dict(cls, dict_data):
+        return cls(
+            dict_data['player_action'],
+            dict_data['player_action'],
+            dict_data['amount'],
+            dict_data['hand_state'],
+            dict_data['action_detail']
+        )
